Flipped bool returned on Command#execute
This may appear odd, but it prevents the need for explicit 'return True' at the end of every command implementation. Now, as long as the execution completes, it is considered to have succeeded - on the other hand, if the execution returns via any ResponseHook method, it is considered to have failed.